portald/requirements.txt
Django~=3.1
django-environ~=0.4.5
cryptography
djangorestframework~=3.11.1
chardet
colorama
cryptography
Markdown
MarkupSafe
mysqlclient~=2.0.1
Pillow~=7.2.0
psutil~=5.7.2
requests~=2.24.0
sqlparse
tabulate
termcolor
urllib3~=1.25.10
ujson~=3.1.0
ruamel.yaml~=0.16.10
pick~=1.0.0
aiohttp~=3.6.2
django-otp~=0.9.4
django-two-factor-auth~=1.12.1
PyGithub~=1.53
jsonschema==3.2.0